# api/api.yaml
openapi: "3.0.4"
info:
  title: "API for the goods coordination system of the GPN LOC."
  version: "0.1.0"
servers:
  - url: "http://localhost:8080/api"
    description: "Local development server."
paths:
  /euroCrate:
    get:
      operationId: getAllEuroCrates
      summary: "Get all euro crates."
      responses:
        "200":
          description: "Successfully retrieved all euro crates."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EuroCrateDto'
    post:
      operationId: createNewEuroCrate
      summary: "Create a new euro crate."
      security:
        - springCookieAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEuroCrateDto'
      responses:
        "201":
          description: "Successfully created a new euro crate."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EuroCrateDto'
        "400":
          description: "Provided euro crate information does not match specifications."
        "401":
          description: "User did not provide valid session details and cannot create a new euro crate."
        "403":
          description: "User does not have permission to create new euro crates."
        "409":
          description: "A euro crate with this combination of name and OC already exists."
  /euroCrate/{id}:
    get:
      operationId: getEuroCrate
      summary: "Get a specific euro crates with all details."
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EuroCrateId"
      responses:
        "200":
          description: "The euroCrate was successfully retrieved."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EuroCrateDto"
        "404":
          description: "The selected euroCrate does not exist."
  /euroCrate/{oc}/{name}:
    get:
      operationId: findEuroCrate
      summary: "Get a specific euro crate with oc and name"
      parameters:
        - name: "oc"
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OperationCenterDto"
        - name: "name"
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "The euroCrate was successfully retrieved."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EuroCrateDto"
        "404":
          description: "The selected euroCrate does not exist."
  /euroCrate/{id}/modify:
    post:
      operationId: modifyEuroCrate
      summary: Modify multiple properties of a crate
      security:
        - springCookieAuth: []
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EuroCrateId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EuroCratePatchDto"
      responses:
        200:
          description: Data updated
        400:
          description: Changes were invalid
        401:
          description: Not logged in
        403:
          description: Not allowed to modify resources
        404:
          description: Crate does not exist
  /euroCrate/setMultipleLocations:
    post:
      operationId: modifyMultipleECLocations
      summary: Modify location of multiple euro crates
      security:
        - springCookieAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  description: Crate IDs
                  items:
                    type: integer
                    format: int64
                location:
                  $ref: "#/components/schemas/LocationDto"
              required:
                - ids
                - location

      responses:
        200:
          description: OK
        400:
          description: Bad request
        401:
          description: Not logged in
        403:
          description: Not allowed to modify resources
        404:
          description: A crate didn't exist
  /euroCrate/{id}/print:
    get:
      operationId: printEuroCrate
      summary: "Generate a printable document for the selected euroCrate."
      security:
        - springCookieAuth: [ ]
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EuroCrateId"
      responses:
        "200":
          description: "Successfully generated the euroPallet document."
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: "Selected euroPallet does not exist."
  /euroCrate/{id}/packingLists:
    get:
      operationId: getPackingListsOfCrate
      summary: Get the associated packing list of the specified crate, if present
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EuroCrateId"
      responses:
        200:
          description: Found the packing list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasicPackingListDto"
        404:
          description: Euro crate does not exist, or no associated packing list
  /euroPallet:
    get:
      operationId: getAllEuroPallets
      summary: "Get all euro pallets."
      responses:
        "200":
          description: "Successfully retrieved all euroPallets."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EuroPalletDto'
    post:
      operationId: createEuroPallet
      summary: "Create a new euro pallet."
      security:
        - springCookieAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEuroPalletDto'
      responses:
        "201":
          description: "Successfully created a new euro pallet."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EuroPalletDto'
        "400":
          description: "Provided pallet information does not match specifications."
        "401":
          description: "User did not provide valid session details and cannot create a new pallet."
        "403":
          description: "User does not have permission to create new pallets."
  /euroPallet/{euroPalletId}:
    get:
      operationId: getEuroPallet
      summary: "Get a specific euro pallet."
      parameters:
        - name: "euroPalletId"
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Successfully retrieved euro pallet."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EuroPalletDto"
        "404":
          description: "Selected euro pallet does not exist."
        "400":
          description: "euroPalletId is not a whole number or doesn't fit in 64 bits"
  /euroPallet/{euroPalletId}/lists:
    get:
      operationId: getEuroPalletLists
      summary: "Get all associated lists of a pallet"
      parameters:
        - name: "euroPalletId"
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Successfully retrieved euro pallet lists"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VeryBasicPackingListDto"
        "404":
          description: "Selected euro pallet does not exist."
        "400":
          description: "euroPalletId is not a whole number or doesn't fit in 64 bits"
  /euroPallet/{euroPalletId}/lastLocation:
    post:
      operationId: updateLastLocationOfEuroPallet
      summary: "Update the last known location of a Euro pallet."
      security:
        - springCookieAuth: [ ]
      parameters:
        - name: "euroPalletId"
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationDto"
      responses:
        "200":
          description: "Successfully updated the last location."
        "401":
          description: "User did not provide valid session details."
        "404":
          description: "Selected euro pallet does not exist."
        "400":
          description: "euroPalletId is not a whole number or doesn't fit in 64 bits, request body is invalid"
  /euroPallet/{euroPalletId}/print:
    get:
      operationId: printEuroPallet
      summary: "Generate a short DIN A5 printable document for the physical pallet."
      security:
        - springCookieAuth: [ ]
      parameters:
        - name: "euroPalletId"
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Successfully generated the pallet document."
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: "Selected pallet does not exist."
        "400":
          description: "euroPalletId is not a whole number or doesn't fit in 64 bits"
...
components:
  securitySchemes:
    springCookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  schemas:
    EuroCrateId:
      type: integer
      format: int64
...
    EuroCratePatchDto:
      allOf:
        - $ref: "#/components/schemas/InformationDto"
        - type: object
          properties:
            deliveryState:
              $ref: "#/components/schemas/DeliveryStateEnumDto"
            location:
              $ref: "#/components/schemas/LocationDto"
...
    NewEuroCrateDto:
      allOf:
        - $ref: "#/components/schemas/InformationDto"
        - $ref: "#/components/schemas/DeliveryStateDto"
        - type: object
          properties:
            operationCenter:
              $ref: "#/components/schemas/OperationCenterDto"
            name:
              description: "A short distinct name describing the contents and/or target location of the euro crate. Has to be unique in combination with the OC."
              type: string
              minLength: 1
            returnBy:
              description: "At which date should the euro crate be returned to the logistics operation center."
              type: string
              format: date
            location:
              $ref: "#/components/schemas/LocationDto"
          required:
            - operationCenter
            - name
            - returnBy
            - location
    EuroCrateDto:
      allOf:
        - $ref: "#/components/schemas/NewEuroCrateDto"
        - type: object
          properties:
            internalId:
              description: 64 bit id describing this crate uniquely
              type: integer
              format: int64
          required:
            - internalId
    NewEuroPalletDto:
      allOf:
        - $ref: "#/components/schemas/InformationDto"
        - type: object
          properties:
            location:
              $ref: "#/components/schemas/LocationDto"
          required:
            - location
    EuroPalletDto:
      allOf:
        - $ref: '#/components/schemas/NewEuroPalletDto'
        - type: object
          properties:
            euroPalletId:
              description: "Unique identifier for the euro pallet."
              type: integer
              format: int64
          required:
            - euroPalletId
...
    VeryBasicPackingListDto:
      allOf:
        - $ref: "#/components/schemas/DeliveryStateDto"
        - type: object
          properties:
            packingListName:
              description: "Human readable name"
              type: string
              minLength: 1
              example: "Lieferung Finanzen"
            packingListId:
              description: "Internal ID for this packing list"
              type: integer
              format: int64
          required:
            - packingListId
            - deliveryState
    BasicPackingListDto:
      allOf:
        - $ref: "#/components/schemas/VeryBasicPackingListDto"
        - type: object
          properties:
            packedOn:
              $ref: "#/components/schemas/EuroPalletDto"
          required:
            - packedOn
...
    DeliveryStateDto:
      type: object
      properties:
        deliveryState:
          $ref: "#/components/schemas/DeliveryStateEnumDto"
      required:
        - deliveryState
    InformationDto:
      type: object
      properties:
        information:
          description: "Free-form text describing the contents of the entity."
          type: string
    DeliveryStateEnumDto:
      type: string
      enum:
        - "Packing"
        - "WaitingForDelivery"
        - "TravelingToGPN"
        - "WaitingAtGPN"
        - "InDelivery"
        - "Delivered"
    LocationDto:
      type: object
      description: "
        locationType indicates what kind of location is present in the DTO.
        Only the relevant field will be not null.
        For example, for the location type 'AtOperationCenter' 
        the field 'operationCenter' is not null.
      "
      properties:
        locationType:
          $ref: "#/components/schemas/LocationTypeDto"
        logisticsLocation:
          $ref: "#/components/schemas/LogisticsLocationDto"
        operationCenter:
          $ref: "#/components/schemas/OperationCenterDto"
        somewhereElse:
          type: string
      required:
        - locationType
    LocationTypeDto:
      type: string
      enum:
        - "Logistics"
        - "AtOperationCenter"
        - "SomewhereElse"
    LogisticsLocationDto:
      type: string
      enum:
        - "Entropia"
        - "InTransport"
        - "LOC"
        - "UnknownAtGPN"
    OperationCenterDto:
      type: string
      enum:
        - "Finanzen"
        - "Backoffice"
        - "Content"
        - "Heralding"
        - "DesignUndMotto"
        - "PresseUndSocialMedia"
        - "LoungeControl"
        - "LoungeTechnik"
        - "Infodesk"
        - "Merchdesk"
        - "Schilder"
        - "Badges"
        - "Trolle"
        - "Kueche"
        - "WOC"
        - "Fruehstueck"
        - "RaumDer1000Namen"
        - "Bar"
        - "Spaeti"
        - "Aussenbar"
        - "Kaffeebar"
        - "Cocktailbar"
        - "NOC"
        - "POC"
        - "VOC"
        - "BuildupAndTeardown"
        - "Infrastruktur"
        - "Deko"
        - "SafeR"
        - "SilentHacking"
        - "Projektleitung"
...